+++ app/DUMMY_DATA.tsx
@@ -1,0 +1,103 @@
+export const perks = [
+  {
+    name: "Free returns",
+    imageUrl:
+      "https://tailwindui.com/img/ecommerce/icons/icon-returns-light.svg",
+    description:
+      "Not what you expected? Place it back in the parcel and attach the pre-paid postage stamp.",
+  },
+  {
+    name: "Same day delivery",
+    imageUrl:
+      "https://tailwindui.com/img/ecommerce/icons/icon-calendar-light.svg",
+    description:
+      "We offer a delivery service that has never been done before. Checkout today and receive your products within hours.",
+  },
+  {
+    name: "All year discount",
+    imageUrl:
+      "https://tailwindui.com/img/ecommerce/icons/icon-gift-card-light.svg",
+    description:
+      'Looking for a deal? You can use the code "ALLYEAR" at checkout and get money off all year round.',
+  },
+  {
+    name: "For the planet",
+    imageUrl:
+      "https://tailwindui.com/img/ecommerce/icons/icon-planet-light.svg",
+    description:
+      "We’ve pledged 1% of sales to the preservation and restoration of the natural environment.",
+  },
+];
+
+export const collections = [
+  {
+    name: "Clothes",
+    href: "#",
+    imageSrc:
+      "https://tailwindui.com/img/ecommerce-images/home-page-04-collection-01.jpg",
+    imageAlt: "Woman wearing a comfortable cotton t-shirt.",
+  },
+  {
+    name: "Drinking",
+    href: "#",
+    imageSrc:
+      "https://tailwindui.com/img/ecommerce-images/home-page-04-collection-02.jpg",
+    imageAlt: "Man wearing a comfortable and casual cotton t-shirt.",
+  },
+  {
+    name: "Home & Living",
+    href: "#",
+    imageSrc:
+      "https://tailwindui.com/img/ecommerce-images/home-page-04-collection-03.jpg",
+    imageAlt:
+      "Person sitting at a wooden desk with paper note organizer, pencil and tablet.",
+  },
+
+  // TODO Implementing UI element to navigate between collections
+  // {
+  //   name: "Phone",
+  //   href: "#",
+  //   imageSrc:
+  //     "https://tailwindui.com/img/ecommerce-images/home-page-04-collection-01.jpg",
+  //   imageAlt:
+  //     "Person sitting at a wooden desk with paper note organizer, pencil and tablet.",
+  // },
+  // {
+  //   name: "Beach",
+  //   href: "#",
+  //   imageSrc:
+  //     "https://tailwindui.com/img/ecommerce-images/home-page-04-collection-02.jpg",
+  //   imageAlt:
+  //     "Person sitting at a wooden desk with paper note organizer, pencil and tablet.",
+  // },
+  // {
+  //   name: "Bags",
+  //   href: "#",
+  //   imageSrc:
+  //     "https://tailwindui.com/img/ecommerce-images/home-page-04-collection-02.jpg",
+  //   imageAlt:
+  //     "Person sitting at a wooden desk with paper note organizer, pencil and tablet.",
+  // },
+  // {
+  //   name: "Covid-19",
+  //   href: "#",
+  //   imageSrc:
+  //     "https://tailwindui.com/img/ecommerce-images/home-page-04-collection-03.jpg",
+  //   imageAlt:
+  //     "Person sitting at a wooden desk with paper note organizer, pencil and tablet.",
+  // },
+];
+
+export const trendingProducts = [
+  {
+    id: 1,
+    name: "Leather Long Wallet",
+    color: "Natural",
+    price: "$75",
+    href: "#",
+    imageSrc:
+      "https://tailwindui.com/img/ecommerce-images/home-page-04-trending-product-02.jpg",
+    imageAlt: "Hand stitched, orange leather long wallet.",
+  },
+  // More products...
+];
